ADD: Refactor for button component
